+++ Changelog.md
@@ -2,6 +2,17 @@
 
 The change log describes what is "Added", "Removed", "Changed" or "Fixed" between each release. 
 
+## 0.2.2
+
+### Added 
+
+- A type property was added to `PrePublishMessage`. 
+- Lots of code style changes
+
+### Fixed
+
+- Bug when configure the Serializer
+
 ## 0.2.1
 
 ### Added